Scale line penalty based on line width
This changes the line penalty computation from a simple

    (target_width - line_width)**2

to a scaled penalty which ranges from 0.0 (a perfect fit) to a
constant INF_PENALTY.

    ((target_width - line_width) / target_width)**2 * INF_PENALTY

Scaling the penalties in this sense avoids the overflows that could
happen in the past when target_width was usize::max_value(). Now the
line cost always goes from 0 to INF_PENALTY, regardless of the actual
target width.

The cost still increases in a quadratic fashion. The INF_PENALTY is
set to 10,000. This makes the new cost computations identical to the
past computations if the target width is 100 columns wide.